Исправил баг
